Move Scalafmt symlink code to SharedCommand
