+++ fastpass/src/main/scala/scala/meta/internal/fastpass/pantsbuild/IntelliJ.scala
@@ -25,7 +25,6 @@ import scala.meta.internal.fastpass.FastpassEnrichments._
 import scala.meta.fastpass.Fastpass
 import scala.meta.internal.fastpass.pantsbuild.commands.SharedCommand
 import scala.meta.internal.fastpass.pantsbuild.commands.SharedOptions
-import scala.meta.internal.fastpass.pantsbuild.commands.SwitchCommand
 
 object IntelliJ {
   def launch(project: Project, open: OpenOptions): Unit = {
@@ -67,7 +66,7 @@ object IntelliJ {
       coursierBinary: Option[Path] = None,
       exportResult: Option[PantsExportResult] = None
   ): Unit = {
-    SwitchCommand.runScalafmtSymlink(project, shared)
+    SharedCommand.runScalafmtSymlink(project, shared)
     val bspJson = project.root.bspJson.toNIO
     Files.createDirectories(bspJson.getParent)
     val coursier = coursierBinary.getOrElse(
